+++ NEWS.md
@@ -8,6 +8,8 @@ to increase sampling efficiency thanks to Andrew Johnson. (#1017, #1053)
 even if the required CDF or quantile functions are unavailable.
 * Update and export `validate_prior`to validate priors supplied by the user.
 * Add support for within-chain threading with `rstan (Stan >= 2.25)` backend.
+* Apply the R2-D2 shrinkage prior to population-level coefficients
+via function `R2D2` to be used in `set_prior`.
 
 ### Bug Fixes